Maxpool forward and backward working
They seem to be quite slow however. Unsure if this is our implementation or not.

+++ trymnist.py
@@ -43,8 +43,8 @@ decay = .0005
 
 m = convnet()
 print("training...")
-train_image_classifier(m, train, batch, iters/2, rate, momentum, decay)
-train_image_classifier(m, train, batch, iters/2, rate*.1, momentum, decay)
+train_image_classifier(m, train, batch, int(iters/2), rate, momentum, decay)
+train_image_classifier(m, train, batch, int(iters/2), rate*.1, momentum, decay)
 print("done")
 print
